refactor: rename solution and project files
