+++ JsonRepairSharp.sln
@@ -1,0 +1,49 @@
+﻿
+Microsoft Visual Studio Solution File, Format Version 12.00
+# Visual Studio Version 17
+VisualStudioVersion = 17.6.33829.357
+MinimumVisualStudioVersion = 10.0.40219.1
+Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "JsonRepairSharp.Tests", "JsonRepairSharp.Tests\JsonRepairSharp.Tests.csproj", "{228E9852-AB2A-4BAF-A822-160BA7AFFF36}"
+EndProject
+Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "JsonRepairSharp-Minimal", "JsonRepairSharp-Minimal\JsonRepairSharp-Minimal.csproj", "{9AB8B1E7-E46A-46D1-BBD4-889D0DCCF862}"
+EndProject
+Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "JsonRepairSharp-CLI", "JsonRepairSharp-CLI\JsonRepairSharp-CLI.csproj", "{0E93551E-2341-4BD2-A8DD-5221F7C58B9D}"
+EndProject
+Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "JsonRepairSharp-GUI", "JsonRepairSharp-GUI\JsonRepairSharp-GUI.csproj", "{6E346FD6-FE34-428C-AE17-9D1B5469E052}"
+EndProject
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "JsonRepairSharp", "JsonRepairSharp\JsonRepairSharp.csproj", "{24A4CD78-2E7D-4A72-8650-395779A75452}"
+EndProject
+Global
+	GlobalSection(SolutionConfigurationPlatforms) = preSolution
+		Debug|Any CPU = Debug|Any CPU
+		Release|Any CPU = Release|Any CPU
+	EndGlobalSection
+	GlobalSection(ProjectConfigurationPlatforms) = postSolution
+		{228E9852-AB2A-4BAF-A822-160BA7AFFF36}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{228E9852-AB2A-4BAF-A822-160BA7AFFF36}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{228E9852-AB2A-4BAF-A822-160BA7AFFF36}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{228E9852-AB2A-4BAF-A822-160BA7AFFF36}.Release|Any CPU.Build.0 = Release|Any CPU
+		{9AB8B1E7-E46A-46D1-BBD4-889D0DCCF862}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{9AB8B1E7-E46A-46D1-BBD4-889D0DCCF862}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{9AB8B1E7-E46A-46D1-BBD4-889D0DCCF862}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{9AB8B1E7-E46A-46D1-BBD4-889D0DCCF862}.Release|Any CPU.Build.0 = Release|Any CPU
+		{0E93551E-2341-4BD2-A8DD-5221F7C58B9D}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{0E93551E-2341-4BD2-A8DD-5221F7C58B9D}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{0E93551E-2341-4BD2-A8DD-5221F7C58B9D}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{0E93551E-2341-4BD2-A8DD-5221F7C58B9D}.Release|Any CPU.Build.0 = Release|Any CPU
+		{6E346FD6-FE34-428C-AE17-9D1B5469E052}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{6E346FD6-FE34-428C-AE17-9D1B5469E052}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{6E346FD6-FE34-428C-AE17-9D1B5469E052}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{6E346FD6-FE34-428C-AE17-9D1B5469E052}.Release|Any CPU.Build.0 = Release|Any CPU
+		{24A4CD78-2E7D-4A72-8650-395779A75452}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{24A4CD78-2E7D-4A72-8650-395779A75452}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{24A4CD78-2E7D-4A72-8650-395779A75452}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{24A4CD78-2E7D-4A72-8650-395779A75452}.Release|Any CPU.Build.0 = Release|Any CPU
+	EndGlobalSection
+	GlobalSection(SolutionProperties) = preSolution
+		HideSolutionNode = FALSE
+	EndGlobalSection
+	GlobalSection(ExtensibilityGlobals) = postSolution
+		SolutionGuid = {8F18053A-AE47-4F0B-A5D9-386AB012E984}
+	EndGlobalSection
+EndGlobal
